flowin-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
anthropic==0.40.0
pydantic[email]==2.9.0
pydantic-settings==2.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
bcrypt==4.2.0
PyJWT==2.9.0
httpx>=0.27.0
